fix: ULog multi info token refactoring, merged with info token
